moved scouting-form-manager to util folder
