+++ src/util/scouting-form-manager.js
